config: support bs=native for failover=election
The bootstrap leader for initializing the replicaset is chosen as the
first candidate instance (because there is no RO/RW information in the
YAML configuration).

Similar to `replication.failover` = `manual` in other regards: once an
instance becomes RW, set it as a bootstrap leader to register new
replicas.

Part of #10857

NO_DOC=to be added in the latest patch of the series
NO_CHANGELOG=same