config: support bs=native for failover=supervised
The `supervised` failover mode assumes an external agent that manages
the replicaset: mainly assigns RO/RW mode. If the bootstrap strategy is
`supervised` or `native`, this agent is also responsible to issue a
command on some instance to bootstrap a replicaset.

This way all the actions related to bootstrap is performed by the
external agent. The only thing that we should do on the instance is to
do nothing about the bootstrap: just wait for the command.

The commit implements exactly this: it disables all the logic that
assigns the RW mode or a bootstrap leader if the following conditions
are met:

* `replication.failover =`supervised` AND
* `replication.bootstrap_strategy` = `supervised` or `native`

Closes #10857

@TarantoolBot document
Title: `replication.bootstrap_strategy` = `native`

The new bootstrap strategy `native` is supported since the 3.4.0
release. It can be enabled using the `replication.bootstrap_strategy`
option in the YAML configuration.

Dev issue: https://github.com/tarantool/tarantool/issues/10857

The new strategy acts more or less similar to the `auto` strategy, but
relaxes its limitations (see details in the linked dev issue).

It is based on the `supervised` strategy and mainly performs two things:

* issues `box.ctl.make_bootstrap_leader({graceful = true})` to bootstrap
  a replicaset
* issues `box.ctl.make_bootstrap_leader()` to keep the bootstrap leader
  record pointing to the instance that is currently in the RW mode (to
  register new replicas)

The choice how to choose a bootstrap leader to initialize the replicaset
depends on the `replication.failover` option. See details in the linked
dev issue.