test/fuzz: add uri tests
The patch adds a couple of tests for the uri Lua module [1].
First one for the `uri.parse()` and `uri.format()` and the
second one for the `uri.escape()` and `uri.unescape()`. The test
`uri_parse_test.lua` is disabled due to bug [2].

The patch also adds an infrastructure for generating CMake tests
for luzer-based tests and removes labels `fuzzing-lua` and
`fuzzing` for non-coverage-guided tests.

1. https://www.tarantool.io/en/doc/latest/reference/reference_lua/uri/
2. https://github.com/tarantool/security/issues/151

Part of #11883

NO_CHANGELOG=testing
NO_DOC=testing