memtx: encapsulate tuple_key_is_excluded check in mvcc
The commit encapsulates call of `tuple_key_is_excluded` in memtx mvcc,
the new function `memtx_tx_tuple_key_is_excluded` should be used instead.
In the next commits, it will be used to encapsulate functional indexes
(functional keys must be passed to `tuple_key_is_excluded` instead of
tuples).

Note that the new helpers must be called only for dirty tuples (we will
use this property later). That's why the commit moves call of
`tuple_key_is_excluded` in `memtx_tx_history_add_insert_stmt` below - we
need to call the new helper when the story for the tuple is created
because that's when it becomes dirty. Calling `index_replace` for
excluded tuples is safe - it's just a no-op actually.

Part of #11100

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring