box: flush WAL queue on sync
`wal_sync()` is expected to flush all prepared txns on disk. Currently
it is not so because we do not flush txns waiting in WAL queue. AFAIU
the issue exists for all the places where we use `wal_sync()` (except
for `box_wait_limbo_acked()` where we already flush WAL queue). The test
is only added for the vinyl index build as there are explicit tickets for
that.

Closes #11118
Closes #11119

NO_DOC=bugfix