relay: fix crash during shut down
In relay there's a message, which travels between relay and tx
threads and which is used to deliver vclock of the relay to tx
thread.

However, when the message is in tx thread, it may happen, that
relay thread dies and tx may try to send message back without
checking, whether the thread is alive.

Let's check `is_paired` flag, which is set to `false`, when
relay thread dies, before sending the cbus message.

Closes #9920

NO_DOC=bugfix