say: do not close stderr at exit
We must not close stderr because it's used for reporting crashes and
memory leaks, but if the logger is configured to log to stderr, it's
closed at exit. Let's fix it.

While we are at it, let's also remove usage of `pm_atomic_store()` and
`pm_atomic_load()` for updating `log::type` because it makes no sense.

The bug was introduced by commit 6d7437dae2dfc ("say: Add several
loggers support").

Close #11323

NO_DOC=bug fix
NO_CHANGELOG=relevant only to developers