memtx: do not use `key_def` along with `cmp_def` in mvcc
In a bunch of functions, related to handling of `count`, we use both
`cmp_def` and `key_def` - one for tuple-with-tuple comparison and
another one for tuple-with-key comparison. Actually, `cmp_def` can be
used for both comparisons, so let's simply drop `key_def` in such places
to simplify the code and future refactoring.

Part of #11100

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring