memtx: encapsulate obtainment of tuple hint in transaction manager
The commit encapsulates obtainment of tuple hint in memtx mvcc - the new
function `memtx_tx_tuple_hint` should be used for that. In the next
commits, it will be used to encapsulate functional indexes (functional
keys are passed as hints to functions working with tuples).

Part of #11100

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring