box: release array and schema in box_insert_arrow
According to the documentation to the Apache Arrow C data interface:

> Consumers MUST call a base structure’s release callback when they
> won’t be using it anymore[^1]

`box_insert_arrow()` is clearly a consumer so it must move or release
the input Arrow array and schema to conform to the specification while
currently it does not. This patch fixes this issue.

Note that for performance reasons (to avoid extra copying), the engine
callback that implements batch insertion should be allowed to move the
input Arrow array and schema to a private location. To allow for that,
we have to move the code that encodes the input Arrow in the IPC format
for xlog from `txn_add_redo()` to `space_execute_insert_arrow()`.

Closes #11338

[^1]: https://arrow.apache.org/docs/format/CDataInterface.html#release-callback-semantics-for-consumers

NO_DOC=bug fix
NO_CHANGELOG=batch insertion isn't implemented in CE