replication: allow bootstrap in RO (supervised st.)
The idea of the `supervised` bootstrap strategy spins around some
external logic/agent/coordinator that makes decisions about a bootstrap
leader.

There are two possible approaches to bootstrap a replicaset:

1. At first, wait for the agent's decision. Do the first `box.cfg()` on
   all the instances after this. The bootstrap leader receives the
   desired configuration plus `read_only = false` (and the command to go
   to bootstrap). All the other replicaset peers receives the desired
   configuration plus `read_only = true`.

   This way we can't use iproto to receive the decision and there are
   different ways to overcome it. For example, cartridge has its own
   implementation of a subset of the iproto protocol to receive such
   pre-box-cfg commands.

   Another possible approach is to implement a separate daemon that
   receives remote commands and feeds them to tarantool admin console.
2. Call `box.cfg()` first with a desired configuration plus `read_only
   = true` on all the peers. Wait for the agent's decision. The chosen
   bootstrap leader receives the command to go to bootstrap plus
   `box.cfg({read_only = false})`.

The second approach feels more native to tarantool and just simpler.
This commit allows it. It is going to be used in the `native` bootstrap
strategy in #10857.

There are alternatives, in fact: see #11286 for some of them. However, I
still think that the given one-line solution worth it (at least until we
have something better designed and implemented).

Needed for #10857

@TarantoolBot document
Title: Bootstrap in RO is allowed in the supervised bootstrap strategy

Now, it is OK to configure all the instances of a replicaset in
read-only (`box.cfg({read_only = true})`) and issue
`box.ctl.make_bootstrap_leader(<...>)` from an external agent on one of
the instances.