replication: wait for bs leader without upstreams
The `supervised` bootstrap strategy allows to run `box.cfg()` with
desired initial configuration first and then issue a command to
bootstrap the database on a particular instance. The instances are
waiting for the bootstrap command during the replication connect timeout
period (30 seconds by default).

However, before this commit this flow doesn't work for a single-instance
replicaset without itself in `box.cfg.replication` (which is fine,
because it is fine to don't have `box.cfg.listen` at all).

This difference requires to implement two different replicaset bootstrap
flows that would significantly complicate a logic in #10857 that
implements the `native` bootstrap strategy.

However, we can eliminate the difference on the lower level and follow
the same logic for the single-instance replicaset as we do for the
multi-instance replicaset. And it magically simplifies the upper level
logic too. Win-win.

This commit makes the initial `box.cfg()` wait for the bootstrap leader
in the `supervised` bootstrap strategy even if `box.cfg.replication` is
`nil` or empty.

Needed for #10857

@TarantoolBot document
Title: box.cfg() waits for a bootstrap leader if there are no upstreams

The `supervised` bootstrap strategy now allows to bootstrap a
single-instance replicaset in the same way as multi-instance one: run
`box.cfg({bootstrap_strategy = 'supervised'})` and wait for the
`box.ctl.make_bootstrap_leader(<...>)` command.