config: add ability to run code before box.cfg
This commit adds 'metadata' support for roles and app scripts.
Currently only 'early_load' tag is supported. If this 'early_load'
tag is set to `true` for a role or a script, this role or script will
be loaded before the execution of `box.cfg` by the config module.

Closes #10182
Closes TNTP-227

@TarantoolBot document
Title: roles and scripts metadata

Metadata can be specified for roles and scripts (module or file in `app`
config section) in lua files. To add metadata to a source file, a
comment must be added to the beginning of the file (or after a shebang).
The metadata comment must contain a YAML-formatted section begging with
'--- #tarantool.metadata.v1' line and ending with '...' line. Example:

```lua
-- This script will be loaded before the first box.cfg() call.
-- --- #tarantool.metadata.v1
-- early_load: true
-- ...
-- More comments can be added here.
```

Currently only 'early_load' tag is supported. Setting 'early_load' tag
makes it so that the tagged file will be loaded (required) before the
execution of `box.cfg` by the config module. During the preload phase
config:get() will not work, both for modules and app scripts.