box: fix session use-after-free
If session is created on demand in fiber we delete it when fiber is
stopped. But we do not clear session and credentials in fiber storage.
It is not an issue for standalone fiber (outside of fiber pool) as fiber
will be destroyed or recycled and will not have chance to execute any
code before that. In case of fiber pool it become dangerous as fiber is
reused. Fortunately we never hit this because we either reset session
and credentials at start in code executed in fiber (like in iproto) or
execute code that does not check access (like in vinyl).

Recently we add `tnt_tx_push()` to execute callback in TX. This time
the issue is revealed.

Let's just clear session and credentials in fiber storage when fiber
is stopped.

Closes #11267

NO_DOC=bugfix

(cherry picked from commit 73af81f8e465efdae5c807c7251a945b4a4e025b)