proofs/tla/wal: introduce the module
This commit implements Write-ahead logging of Tarantool. It's
the high-level abstraction, which does processing of the cbus
messages from the Tx thread to the Wal. It assigns LSNs to the
received xrow entries and adds them to the wal.rows, which
implements disk. The write to disk is atomic and non-failing
for now. After that the written transaction is sent back to Tx
in order to do all needed on Wal write callbacks.

Part of #5206

NO_CHANGELOG=proofs
NO_TEST=added
NO_DOC=later