memtx: add tuple hint to snapshot cleaner entry
Recently, support of functional indexes was added to memtx MVCC - the
implementation is based on encapsulation of tuple hint because
functional indexes store keys of tuples as hints actually. In order to
support the feature for read-view, we should save tuple hint to snapshot
cleaner entry - that's what the commit does.

Follow up #11099

NO_TEST=in ee
NO_CHANGELOG=follow-up
NO_DOC=follow-up