memtx: do not use the same variable for result and successor
Recently, we started to return successor from single-key functional
indexes, but unfortunately memtx store result and successor of replace
in the same variable, and replace of functional index has an assertion
checking if `result` stores an expected value - and it fails when
successor overwrites it. The commit fixes the problem and adds a note to
`index_replace` description to avoid such surprises in future.

The test is simple but we have no tests related to functional indexes
(at least, without MVCC) writtent in luatest, so let's simply extend the
old diff test - add a bunch of replaces to the basic case.

Follow up #11099

NO_CHANGELOG=bug is not released
NO_DOC=bugfix