memtx: use index definition's `key_def` for point hole comparisons
Currently, we assume that point hole equivalence classes are defined by
their keys' binary (MsgPack) representation, and use `memcmp` for
comparison. However, this is not true in the case of numeric key
definitions and collations.

To fix this, let's use the corresponding index's key definition for
comparison. Storing the key length then becomes redundant, so let's drop
it.

Closes #10159
Closes #11292

NO_DOC=<bugfix>