engine: abort writing transactions after ro switch
The `switch_to_ro` callback was first introduced in the `engine_vtab`
as a solution to the issue in vinyl, where it was possible to commit
a transaction after becoming read-only. Since mvcc was introduced in
the memtx engine, a similar issue became possible. A possible
solution in the memtx engine with implementing the `switch_to_ro`
callback would just abort all writing transactions related to memtx,
so cross-engine transactions would be aborted in vinyl and memtx,
which doesn't introduce a bug but seems redundant. So now we just
abort all writing transactions disregarding the engine which resulted
in the `switch_to_ro` callback becoming useless. It is now impossible
for a fiber to awake and commit after switching the node to read-only.
Also more explicit error messages were implemented to address this
situation.

Fixes #9937
NO_DOC=bugfix

(cherry picked from commit 84546f9639b69499b974779bda66f358e6ce8108)