core: store evio service entries in an rlist
Refactor the `struct evio_service` so it contains an rlist of entries
which doesn't need to have some fixed size after the creation. The main
reason for the change is that we don't know the number of bounded
entries for each uri on service creation once #7152 being fixed.

Part of #7152

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring