memtx: introduce the O(n) SK build
A new entity introduced: memtx sort data file. This is a binary file
with text header (like XLOG) containing the data required to perform
recovery using the O(n) SK sorting algorithm. The file has `.sortdata`
extension and garbage collected the same way as regular `.snap` files.

The new `memtx_sort_data_enabled` config option controls whether to
write the sort data file on `box.snapshot()` or use it during recovery.
The option can be changed dynamically, so user can write checkpoints
with or without the new data regardless of had the data been used on
current instance recovery or had not.

But the sort data can only be used if:
- The `_index` space has no `before_replace` triggers (so we avoid
  the sort data invalidation of an index because the index had been
  altered by a `before_replace` trigger).
- The space isn't a system space (these generally don't benefit from
  the new SK sorting algorithm).

Of course, the new approach is also not used during forced recovery.

@TarantoolBot document
Title: Document the MemTX's new secondary key build approach
Since: 3.5

Now it's possible to sort secondary keys using a new O(n) sorting
algorithm using an additional data written into the snapshot. The
feature can be enabled with a new `memtx_sort_data_enabled` option
in `box.cfg` or `memtx.sort_data_enabled` in the instance config.

With the option set to `true`, additional data is saved into a
separated `.sortdata` file on snapshot creation and the data is
used on recovery. The default value is `false`: the behavior has
to be turned on by the user if required.

The option is changeable in runtime. So one can, for example, enable
it on recovery to use the new SK sorting approach, but disable it
before creating a new snapshot, so only `.snap` file will be created
and the O(n) SK sort will not be available for the snapshot.

The performance impact of the option depends on the persistent
storage read/write speed and the amount of tuples and secondary
keys in spaces (the more the better the new approach is). Also,
the approach only utilizes a single CPU core.

As a downside, it creates an additional footprint overhead on
recovery time (up to ~45 bytes per tuple) and overhead during
the snapshot creation (since we have to write the sort data to
the persistent storage).