perf/tools: generate a benchmark config
The patch updates a script `aggregate.lua` so that it can be able
to generate a JSON config with benchmark names and their subtests.

The config will be created in a file
<build_dir>/perf/output/config_bench.json by the command
`make test-perf-aggregate`.

The patch is needed for the following commits.

NO_CHANGELOG=performance
NO_DOC=performance
NO_TEST=performance