core: bind to all addresses
Bind all addresses returned by `getaddrinfo` except for duplicates
caused by glibc behavior which is the case on the CI machine and may be
the result of multiple `localhost` entries in `/etc/hosts`. Also bind
all ipv6 results with the `IPV6_V6ONLY` option so it is possible to
listen ipv4 and ipv6 services on a single port number.

Fixes #7152

@TarantoolBot document
Title: bind all URIs matching the `listen` box configuration parameters.