limbo: speed up synchro confirm writing
In the case of a single node cluster, the limbo worker was not writing
synchro confirms fast enough.

The problem was, firstly, that sometimes it was possible to queue a
confirm writing immediately (without yielding) from the
`txn_on_journal_write` write-callback, instead of waking up the limbo
worker to queue this writing on the next iteration of the event loop.
The second problem was that limbo worker wrote confirms synchronously.
There is no point in waiting for the confirm writing to finish, because
while the worker is waiting, `volatile_confirmed_lsn` could have bumped
and it doesn't matter how the previous writing ends, because you can
already write a confirm to the updated greater lsn and retry only it if
necessary, there is no point in retry unsuccessful writes with lower lsn.

In the 1mops_write bench with parameters --fibers=6000 --ops=1000000
--transaction=1 --sync, performance increases by almost 2 times on a single
node cluster.

Closes #11404

NO_DOC=performance improvement
NO_TEST=performance improvement