core: introduce the pointer to pointer hash
Pointers can have different alignment. Let's ignore the last few bits
hashing the table entries, so the pointer alignment is assumed to be
16 bytes, which is the minimal size of a tuple with a single field
aligned to the default small allocator granularity, which is 8. It's
also the malloc'd pointer alignment on 64-bit glibc, so it must not
hurt the `mh_ptr` use in MemTX (malloc'd extents container).

The container is to be used in the future for storing tuple pointer
associations during recovery using the O(n) SK sorting algorithm.

Needed for #10847

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal