test: fix flakiness of the election + qsync stress test
The source of flakiness of the test came from its sensitivity to the
replication timeout, and, consequently, the leader death timeout: the
leader could get re-elected while we were trying to send an insert request
to the old leader — we would never notice the ER_READONLY error. To fix
this, let's increase the replication timeout to 1 second (the default
value), which seems to give enough stability.

While we are here, port the test to luatest to simplify its maintenance
with the following principal changes:
1. After a new leader gets elected, check the emptiness of his limbo rather
   than writing an asynchronous transaction.
2. After the old leader gets restarted, wait for cluster synchronization
   with him.

Closes #10543

NO_CHANGELOG=<flaky test>
NO_DOC=<flaky test>