core: make decimal hold 76 digits of precision
We are going to add decimal256 support to Tarantool. This type will have
256 bits of precision which corresponds to 76 demical digits. So to be
able to pass decimal values to field of this type we need to increase
precision of our decimal from 38 to 76 digits as well.

This does not change MsgPack representation of decimals with 38 and less
significant digits. Also thanx to deliberately chosen size of
`box_decimal_t' in the commit 5c1bc3dacbe6 ("decimal: add the library
into the module API") we do not break public decimal ABI.

Part of tarantool/tarantool-ee#1188

NO_DOC=later
NO_CHANGELOG=later