perf: drop column_scan_module.init
We're going to move the code that creates the test data set to the C
module. Since this is done before we open a read view in 'init', let's
drop the 'init' function and initialize a read view on demand to avoid
confusion ('init' is usually called first).

NO_DOC=perf test
NO_TEST=perf test
NO_CHANGELOG=perf test