config: add `fail_if` tag for roles and scripts
Closes #10987
Closes TNTP-1016

@TarantoolBot document
Title: `fail_if` tag for roles and scripts.

If the tag `fail_if` is set to an expression string role/script
loading will raise an error, if fail_if expression is evaluated
to be `true`.

The syntax for `fail_if` tag is the same as in config `conditional.if`
section.

Example:
```lua
-- --- #tarantool.metadata.v1
-- fail_if: "tarantool_version < 3.1.0"
-- ...
```