core: move EE error injections to extras
To introduce a new error injection to the EE repository, one has to
first add it to the CE repository and bump the CE submodule in EE.
This is inconvenient. Let's move all EE error injections to extras
and define them in the EE repository.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring