ci: add Tarantool's revision in reusable workflows
There is no way to detect the revision of the Tarantool or Tarantool's
workflow, which is called in the role of a reusable workflow, since all
context is determined by the caller [1]. When a reusable workflow is
called on the non-master Tarantool branch, it is desirable to check out
Tarantool to the corresponding branch. For now this branch is hardcoded
as master, and it is easy to forget to update the hardcoded value after
creating a new branch for the Tarantool's release. OTOH, the caller
should already update the target branch in its own workflow, so it is
easier to maintain this logic in one place (the caller's workflow).
Thus, this patch adds an additional optional parameter -- revision of
the Tarantool to be checked out during the integration. It is set to
'master' by default to avoid breaking existing workflow integration.

[1]: https://docs.github.com/en/actions/sharing-automations/reusing-workflows

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci

(cherry picked from commit 5465e559074ac7431d83964fdbe24a29a0eaeae1)