xlog: drop xdir_type
The xdir_type passed to xdir_create() and stored in the xdir struct is
used for several purposes:
 1. To initialize xdir filetype and filename_ext.
 2. To enable force recovery for the WAL directory.
 3. To determine whether we need to store the previous vclock in meta
    (we do this for WAL files to check for gaps on recovery).
 4. To fail assertion if xdir_touch_xlog() is called for a file that
    isn't a snapshot.

P.1 can be fixed by passing a string instead of a type.

P.2 isn't really needed because we set the force_flag after initializing
the xdir. We just need to do that for the WAL dir as well.

P.3 can be fixed by adding a special flag to the xdir struct, similar
to the force_recovery flag.

P.4 isn't really needed and can be dropped.

Let's drop xdir_type so that we don't need to register a new file type
in the CE repository whenever we want to use xdir in EE.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring