box: add internal box_on_insert_arrow trigger
The trigger is invoked on an arrow insertion. The trigger callback has
passed a context with space, the transaction the insertion is issued
in, the data array, schema and field mapping. The trigger is run only
if the batch validation is passed. It's going to be used for auditing
INSERT_ARROW events in the EE version.

Part of tarantool/tarantool-ee#1174

NO_DOC=ee
NO_TEST=internal
NO_CHANGELOG=internal