memtx: disallow alter of primary index with dependent secondary
Rebuild of several indexes in one statement is broken because we need
to maintain consistency of built indexes while building the next ones,
and we don't do that. Such rebuilds can happen only when primary index
is altered and all non-unique and nullable indexes need to be rebuilt as
well. Let's simply forbid such alters.

We could forbid any alter of primary index, and the patch would be even
a bit simpler, but let's allow to alter primary in a space without
dependent secondary indexes - it seems to work fine, and people would
use it, at least someone reported this problem, hence, someone actually
tried to alter primary.

Note that even if the space is empty, alter of primary with dependent
indexes will be forbidden - otherwise, the patch could be much more
complicated.

Closes #10951

NO_DOC=bugfix