schema: introduce new system space filters
The commit populates schema with the new system space `_filters`. This
space will be used to define "filters" - aggregates collected by index
in order to be used on future reads for data-skipping.

In this commit, we only populate schema with the new system space. The
system trigger for this space, containing actual DDL logic, will be
introduced in the next commit.