box: validate arrow batches prior to space call
This is done in order to check the batch array early and receive the
arrow array to field mapping to be used in batched insertion trigger
(the trigger is to be introduced in the next commit).

The modified tests expected ER_UNSUPPORTED, but, since the batch
validation based on the space format is done prior to calling the
engine arrow insertion methods now, the error is now on the wrong
space format. So update the tests accordingly.

Region allocator does not support size == 0, the field map allocation
and early validation return crutches introduced.

Needed for tarantool/tarantool-ee#1174

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal