box: implement DDL for filters
The commit implements a trigger for system space `_filters` that
processes DDL for filters. Note that the trigger calls stubs in CE
version so any attempt to declare a filter will throw `ER_UNSUPPORTED`
error. Also, the commit adds a new method `alter_filters` to Lua index
objects to allow user to do this DDL conviniently.

When a space or index is dropped, all its filters are dropped as well.
If one will try to drop an index having a filter through the system
space `_index`, an error will be thrown since such drop is forbidden.
When index is being rebuilt, engine must handle `filters` from the old
index by itself - an assertion after rebuild checks if it is handled.