memtx: do not record full count in MVCC multiple times
We used to create `GAP_COUNT` items on each index count with empty
key performed by an txn. The thing is, it's not required to register
such an item for a given transaction more than once.

We're going to record `index:len()` calls in the MVCC in the future,
so let's prevent possible regressions by only registering full count
events once for a transaction. The simplest solution - just insert
full count gap items in the end of the index gap list, so we can
perform a faster check if such an item had been created (and thus,
a full count had been performed) by a given transaction and skip a
new item creation.

Needed for #10149

NO_DOC=internal
NO_CHANGELOG=internal