raft: fix box.ctl.promote hang
Before this patch `box.ctl.promote` could lead to hang. It was happening in
the situation when one candidate server during the promote got a message
from follower that leader was already seen.

This particular scenario can reproduce `box.ctl.promote` hang:
1) We have a replicaset with 3 servers.
2) `server1` loses an upstream connection with `server3`.
3) Then we call `box.ctl.promote` on `server1`. The term of `server1`
is higher than term of other servers.
4) When `server1` sends broadcast messages to other servers, `server3`
starts election process due to election timeout and becomes a leader. Now,
terms of all servers are equal.
5) During the promote process, `server1` gets a message from follower
node - `server2` that leader - `server3` was already seen.

The hang is happened because the raft state machine does not handle
a scenario when the candidate server during election process gets a
message from follower that leader was already seen. It is important
to note that candidate server should not have a connection to newly
elected leader. As a result the candidate server continues to
believe that there is no leader and sends election broadcast
messages to other servers.

Since this kind of promote does not disrupt the cluster it is
decided not to change a low level raft implementation. In order to
avoid a hang in user's fiber we introduce a timeout for `box.ctl.promote`,
after which the control is transferred back to the user. The timeout for
promote is set to `election_timeout * 1.1`. At the end of this time the
`TimedOut` error will be raised.

Closes tarantool#10836

NO_DOC=bugfix