fuzz: support nullable memcs indexes in engine fuzzing
When the engine fuzzer was adapted for `memcs` engine (7b36cb4),
nullable indexes were not supported by this engine, so we had
to add a branch to prevent creation of nullable indexes for
`memcs`. That's not the case now, so the commit removed that
branch.

Also, `memcs` is the only engine that doesn't support creation of
non-nullable indexes over nullable fields - the commit makes fuzzer
handle this case. Before this change, fuzzer for `memcs` could fail
right at the start because it could try to create such index during
setup.

Follows up tarantool/tarantool-ee#834

NO_CHANGELOG=fuzz
NO_DOC=fuzz