third_party: bump nanoarrow
This is required to support decimal32 and decimal64 format in Arrow
in EE.

This is bump on master. Currently we are on non release version also.
Master do not have issues fixed by commits in our fork so we can use
just master branch.

We also force lib install dir to "lib". On some systems it is "lib64"
by default and we assume it is always "lib".

Part of tarantool/tarantool-ee#1188

NO_TEST=submodule bump
NO_CHANGELOG=submodule bump
NO_DOC=submodule bump