limbo: handle spurious wakeups on space waiting
The limbo has a size limit. Only the entries of total sum size not
exceeding the limit are allowed to be persisted, replicated, and
waiting for acks at any time.

The entries beyond that limit must stay volatile and wait for
space to appear.

The difference between volatile and submitted entries is that the
volatile ones can be rolled back without any action in WAL. Since
they aren't written there anyway.

There was a bug in the waiting for limbo space. The waiters
wouldn't form any queue with a specific order. As a result with
some spurious wakeups it was possible to do box.commit() of
different transactions in one order, but get them actually written
to journal and committed in another order.

Another issue but very closely related was that if some txns go
to wait for the limbo space, and the non-last one is rolled back
(due to fiber cancellation, for example), then it wouldn't
cascading rollback the newer txns. This could result in data being
corrupted even locally, without any replication.

And the third issue is coming directly out of the second one -
the rollbacks of the txns waiting for the limbo space were not
visible anywhere. As a result it was possible on join of a new
replica to create a read-view with some data from a txn waiting
for space in the limbo, roll that txn back, and the join-process
wouldn't notice. Then the replica would get a snapshot with already
rolled back data.

The patch addresses that:
- All entries go to the limbo, regardless of the limbo's current
    size.
- The volatile entries, when don't fit into the limbo by size,
    remain volatile and block the current fiber. Those entries'
    sizes aren't accounted in the limbo.
- When a volatile entry is rolled back, it directly has access to
    all the newer volatile entries too, and cascade-rolls them
    back as well.
- Since even the volatile txns are in the limbo, all the code
    trying to "wait for limbo confirm" automatically sees their
    commits and rollbacks.

Part of #11180

NO_DOC=bugfix

(cherry picked from commit 20aad159d065bbd6e5d36bc465af8adb2bc59c30)