limbo: handle journal's cascading rollback
When the txns sent to the journal are rolling back, they used to
ignore any newer txns which might also be prepared but not sent to
the journal yet.

Those were the limbo txns which are waiting for space. I.e. they
aren't even submitted to the journal at all, and aren't visible to
it. They wouldn't be rolled back. And that could lead to data
inconsistency.

Lets notify the limbo about a cascading rollback happening. So it
could rollback all the txns which aren't visible to the journal.
And then the journal would rollback its own volatile txns. And
then the txns which got a rollback inside the WAL thread.

Not all the steps are always done though. For example, if a
volatile journal entry waiting for space was cancelled (like when
its fiber-owner is cancelled), then the cascading rollback will
kill everything newer than this entry + this entry. But won't
touch the older entries which could also be volatile or be already
sent to WAL for writing.

See how the limbo notification is done via a callback (a function
pointer). This is the way to avoid the direct journal->limbo
dependency in the code.

Part of #11180

NO_DOC=bugfix
NO_CHANGELOG=already written

(cherry picked from commit c656f58d8fb37b4c13de5e520dccafb569c92ebf)