memtx: allow to bypass tuple size limitation internally
Option `memtx_max_tuple_size` can be useful for a user to control the
workload, but when we want to allocate a huge tuple internally, we
shouldn't take this limitation into account. So let's populate exported
`memtx_tuple_new_raw` function with a new argument `limit_size` - tuple
size limitation works only when it is passed as `true`.