config: reload config by SIGUSR2
The idea is quite straightforward: once tarantool receives `SIGUSR2`, it
reloads the YAML configuration. If the YAML configuration is not
provided (the instance is started using the 2.x Lua script flow), it is
no-op.

Consequent signals may be coalesced by the OS if they're sent fast
enough. However, if tarantool is in processing of one reload triggered
by a signal, the next signal is queued and a second reload happens after
the first one.

The current implementation has a nuance: `config:reload()` called from
Lua is not queued. If `SIGUSR2` arrives during `config:reload()`, a log
warning issued and the second reload doesn't happen. Similarly,
automatic reload from a remote configuration source (etcd/config
storage) is not queued.

These nuances are *not* an API guarantee and I hope that we'll process
various kinds of reload requests in some uniform way in a future.

Fixes #11585

@TarantoolBot document
Title: config: reload configuration by SIGUSR2

Now, tarantool reloads the YAML configuration if the `SIGUSR2` signal
arrives.