box: fix crash while nil-uuid subscribe request
Before this patch a master node could crash when it processed an
anonymous iproto subscribe request which had a nil instance uuid.
The reason of this error is that in `box_connect_replica` we didn't
throw an error when non-exist replica tried to connect to node. It
led to a situation when the nullified replica object was tried to
dereferenced while iproto subscribe request. Now, we raise an
ER_NIL_UUID error if we get an iproto request with nil `instance_uuid`.

After this change `test_fetch_snapshot_no_uuid` and
`test_checkpoint_join` start fail. The reason of these failures is an
incorrect err.type and err.msg of iproto response when we try to send
an iproto request with nil `instance_uuid`. Now, these tests are fixed
by changing an err.type and err.msg to appropriate values.

Also we change `gh_10155_make_iproto_resistant_to_misusage` test because
`write_fetch_snapshot` wasn't be able to send iproto fetch_snapshot
request with uuid. It led to situation when ER_NIL_UUID error was raised
during fetch_snapshot and as a result we weren't be able to read
snapshot after this. It broke
`test_iproto_crash_fetch_snapshot_subscribe` and
`test_iproto_crash_fetch_snapshot_subscribe_not_anon`.

Closes #11531

NO_DOC=bugfix